requirements.txt
pyserial
astropy
dataclasses;python_version=="3.6"
pyjwt[crypto]
requests
RPi.GPIO